requirements.txt
apify-client==1.8.1
apache-airflow==2.10.3
pandas==2.2.2
